normalization methods moved to common/utils
