+++ common/utils/__init__.py
@@ -1,0 +1,11 @@
+from .money import format_money
+from .normalization import (
+    normalize_digits,
+    normalize_email,
+)
+
+__all__ = [
+    'format_money',
+    'normalize_digits',
+    'normalize_email',
+]
