+++ users/helpers/__init__.py
@@ -1,14 +1,12 @@
 from .auth import (
     generate_username,
     issue_tokens_for_user,
-    normalize_email,
     run_actions_after_authentication,
     set_unusable_password,
 )
 from .profiles import ensure_listener_profile
 
 __all__ = [
-    'normalize_email',
     'ensure_listener_profile',
     'issue_tokens_for_user',
     'generate_username',
