+++ users/helpers/auth.py
@@ -4,6 +4,8 @@ import logging
 
 from rest_framework_simplejwt.tokens import RefreshToken
 
+from common.utils.normalization import normalize_email
+
 from store.services.cart_service import CartService
 
 logger = logging.getLogger(__name__)
@@ -17,11 +19,6 @@ def issue_tokens_for_user(user) -> dict:
         'access': str(access_token),
         'refresh': str(refresh_token),
     }
-
-
-def normalize_email(email: str) -> str:
-    """Нормализация email."""
-    return email.strip().lower()
 
 
 def generate_username(email: str, attempt: int) -> str:
