+++ users/models/core_user.py
@@ -4,6 +4,8 @@ from django.contrib.auth.models import AbstractUser
 from django.db import models
 from phonenumber_field.modelfields import PhoneNumberField
 
+from common.utils import normalize_email
+
 
 class CoreUser(AbstractUser):
     """Кастомная модель пользователя.
@@ -37,7 +39,7 @@ class CoreUser(AbstractUser):
 
     def save(self, *args, **kwargs):
         if self.email:
-            self.email = self.email.strip().lower()
+            self.email = normalize_email(self.email)
         super().save(*args, **kwargs)
 
     class Meta:
